pyproject.toml
[tool.poetry]
name = "analysis-of-banking-operations"
version = "0.1.0"
description = ""
authors = ["Nika Alexandrovna <megrelynika@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
requests = "^2.32.4"
poetry-core = "^2.1.3"
pandas = "^2.3.0"
openpyxl = "^3.1.5"
python-dotenv = "^1.1.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.16.1"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-mock = "^3.14.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 119
exclude = '''
/(
    \.git
   | venv
)/
'''

[tool.isort]
line_length = 119
profile = "black"
skip_glob = ["venv/*"]

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = "venv"
